Directory refactor: Removed unnecessary directory nesting from test/
