+++ tests/integration/error/JsonErrorHandlerTest.php
