+++ tests/unit/ErrorToExceptionTest.php
