+++ tests/unit/ShutdownHandlerTest.php
