+++ tests/unit/error/JsonErrorHandlerTest.php
