+++ tests/unit/middleware/SlotTest.php
