+++ tests/unit/routing/SingleControllerStrategyTest.php
