+++ tests/unit/routing/StaticStrategyTest.php
